# docs/swagger.yaml
basePath: /
definitions:
...
  models.CinemaSeat:
    properties:
      category:
        type: string
      is_available:
        type: boolean
      price:
        type: number
      seat_number:
        type: string
    type: object
  models.GetCinema:
    properties:
      city:
        type: string
      date:
        type: string
      id:
        type: integer
      name:
        type: string
      time:
        type: string
    type: object
...
  models.Ticket:
    properties:
      date:
        description: 'Format: 07 Jul 2025'
        type: string
      movie_title:
        type: string
      quantity:
        type: integer
      seats:
        description: 'Example: ["A1", "A2", "A3"]'
        items:
          type: string
        type: array
      time:
        description: 'Format: 2:00 PM'
        type: string
      total_price:
        type: number
    type: object
...
  pkg.PageInfo:
    properties:
      currentPage:
        type: integer
      nextPage:
        type: integer
      prevPage:
        type: integer
      totalData:
        type: integer
      totalPage:
        type: integer
    type: object
  pkg.Response:
    properties:
      error: {}
      message:
        type: string
      pageInfo:
        $ref: '#/definitions/pkg.PageInfo'
      result: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /order/cinemas:
    get:
      consumes:
      - multipart/form-data
      description: Retrieve a list of cinemas based on the movie ID, location, date,
        and time.
      parameters:
      - description: Movie ID
        in: query
        name: movie_id
        required: true
        type: integer
      - description: Location
        in: query
        name: location
        required: true
        type: string
      - description: 'Date (format: YYYY-MM-DD)'
        in: query
        name: date
        required: true
        type: string
      - description: 'Time (format: HH:mm:ss)'
        in: query
        name: time
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.GetCinema'
                  type: array
              type: object
        "400":
          description: Invalid input
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Find cinemas by movie and location
      tags:
      - Cinemas
  /order/payments:
    post:
      consumes:
      - multipart/form-data
      description: Process a payment for a given transaction ID and update its status.
      parameters:
      - description: Transaction ID
        in: formData
        name: transaction_id
        required: true
        type: integer
      - description: Payment Method
        in: formData
        name: payment_method
        required: true
        type: string
      - description: VA Number
        in: formData
        name: virtual_account_number
        required: true
        type: string
      - description: Ammount
        in: formData
        name: amount
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment processed successfully
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid input
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Process a payment for a transaction
      tags:
      - Transactions
  /order/seats:
    post:
      consumes:
      - multipart/form-data
      description: Get a list of available seats for a specific movie and cinema schedule.
      parameters:
      - description: Movie ID
        in: formData
        name: movie_id
        required: true
        type: integer
      - description: Cinema ID
        in: formData
        name: cinema_id
        required: true
        type: integer
      - description: 'Date (format: YYYY-MM-DD)'
        in: formData
        name: date
        required: true
        type: string
      - description: 'Time (format: HH:mm:ss)'
        in: formData
        name: time
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CinemaSeat'
                  type: array
              type: object
        "400":
          description: Invalid input
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Retrieve available seats for a cinema
      tags:
      - Seats
  /order/ticket:
    get:
      consumes:
      - multipart/form-data
      description: Get ticket information based on a transaction ID.
      parameters:
      - description: Transaction ID
        in: query
        name: transaction_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ticket details retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Ticket'
              type: object
        "400":
          description: Invalid input
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Retrieve ticket details for a transaction
      tags:
      - Tickets
  /order/transactions:
    post:
      consumes:
      - multipart/form-data
      description: Process a payment for a given transaction ID and update its status.
      parameters:
      - description: User ID
        in: formData
        name: user_id
        required: true
        type: integer
      - description: Movie ID
        in: formData
        name: movie_id
        required: true
        type: integer
      - description: Cinema ID
        in: formData
        name: cinema_id
        required: true
        type: integer
      - description: 'Date (format: YYYY-MM-DD)'
        in: formData
        name: date
        required: true
        type: string
      - description: 'Time (format: HH:mm:ss)'
        in: formData
        name: time
        required: true
        type: string
      - description: Location
        in: formData
        name: location
        required: true
        type: string
      - collectionFormat: multi
        description: Seat Numbers
        in: formData
        items:
          type: string
        name: seat_numbers[]
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Transaction created successfully
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid input
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Process a payment for a transaction
      tags:
      - Transactions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"